added a scheme to generate a copy number for ecal detectors. This scheme is still dependent on how the geometry is setup in the GDML file. Therefore, not a ideal choice. To give more detail about the issue,
Due to the geometrical design of the ecal, the copy id is not unique. This is due to one fraction of the ecal is first generated and then it is repeated along the X-Y plane. The copy id for ecal blocks within the first generated fraction (ecal_block_id) is unique. This id is then repeated in the other ecal blocks  (ecal_id). Therefore, ecal_id and ecal_block_id are both required to extract an unique id (copyID). In each ecal block there are 389 volumes with unique ids. There are 3461 ecal blocks in the ecal detector.