Added hit time to the remoll generic hit detectors (remollGenericDetectorHit). The parameter hit.t in the TTree now stores the global hit time (time since the event was created).
Two timing parameters added to the ecal detector.

Global time of the first hit on an ecal block : cal.st[cal.n]
Global time of the first hit on an ecal block : cal.et[cal.n]

The timing parameter for the generic detectors is tested and appear to be working. I still have to verify the global time of the first and last hit on an ecal block. So proceed with care